+++ packages/terra-graphs-docs/CHANGELOG.md
@@ -23,3 +23,4 @@
 * Changed 
   * Moved core documentation to terra-dev-site.
   * Updated Bubble graph documentation and examples to be carbon centric.
+  * Updated Line graph documentation and examples to be carbon-centric.