+++ packages/terra-graphs/CHANGELOG.md
@@ -21,6 +21,6 @@
   * Added reflow to the Scatter graph component.
 
 * Changed
-  * Made changes to WDIO test url's according to terra rules.
+  * Made changes to WDIO test url's according to terra rules. 
 
 * Initial stable release